ef35a38ad29c@2024-04-07_09-03-46: fix ape-e with fsdp

+++ ape/evaluation/oideval.py
@@ -53,7 +53,7 @@ def compute_average_precision(precision, recall):
 
     if not isinstance(precision, np.ndarray) or not isinstance(recall, np.ndarray):
         raise ValueError("precision and recall must be numpy array")
-    if precision.dtype != np.float or recall.dtype != np.float:
+    if precision.dtype != float or recall.dtype != float:
         raise ValueError("input must be float numpy array.")
     if len(precision) != len(recall):
         raise ValueError("precision and recall must be of the same size.")
@@ -448,8 +448,8 @@ class OIDEval:
                 tps = np.logical_and(dt_m, np.logical_not(dt_ig))
                 fps = np.logical_and(np.logical_not(dt_m), np.logical_not(dt_ig))
 
-                tp_sum = np.cumsum(tps, axis=1).astype(dtype=np.float)
-                fp_sum = np.cumsum(fps, axis=1).astype(dtype=np.float)
+                tp_sum = np.cumsum(tps, axis=1).astype(dtype=float)
+                fp_sum = np.cumsum(fps, axis=1).astype(dtype=float)
 
                 dt_pointers[cat_idx][area_idx] = {
                     "tps": tps,
@@ -479,7 +479,7 @@ class OIDEval:
                             pr[i - 1] = pr[i]
 
                     mAP = compute_average_precision(
-                        np.array(pr, np.float).reshape(-1), np.array(rc, np.float).reshape(-1)
+                        np.array(pr, float).reshape(-1), np.array(rc, float).reshape(-1)
                     )
                     precision[iou_thr_idx, :, cat_idx, area_idx] = mAP
